Struct criterion_stats::bivariate::regression::Slope

pub struct Slope<A> where A: Floaty(pub A);

A straight line that passes through the origin y = m * x

Methods

impl<A> Slope<A> where A: Floaty

fn fit(data: Data<A, A>) -> Slope<A>

Fits the data to a straight line that passes through the origin using ordinary least squares

Time: O(length)

fn r_squared(&self, data: Data<A, A>) -> A

Computes the goodness of fit (coefficient of determination) for this data set

Time: O(length)