Struct criterion_stats::bivariate::Data

pub struct Data<'a, X, Y> where X: 'a, Y: 'a(_, _);

Bivariate (X, Y) data

Invariants:

No NaNs in the data

At least two data points in the set

Methods

impl<'a, X, Y> Data<'a, X, Y>

fn len(&self) -> usize

Returns the length of the data set

fn is_empty(&self) -> bool

Checks whether the data set is empty

fn iter(&self) -> Pairs<'a, X, Y>

Iterate over the data set

impl<'a, X, Y> Data<'a, X, Y> where X: Floaty, Y: Floaty

fn new(xs: &'a [X], ys: &'a [Y]) -> Data<'a, X, Y>

Creates a new data set from two existing slices

fn bootstrap<T, S>(&self, nresamples: usize, statistic: S) -> T::Distributions where S: Fn(Data<X, Y>) -> T, S: Sync, T: Tuple, T::Distributions: Send

Returns the bootstrap distributions of the parameters estimated by the statistic

Multi-threaded

Time: O(nresamples)

Memory: O(nresamples)

fn x(&self) -> &'a Sample<X>

Returns a view into the X data

fn y(&self) -> &'a Sample<Y>

Returns a view into the Y data