Trait criterion_stats::tuple::TupledDistributions

pub trait TupledDistributions: Sized {
    type Item: Tuple<Distributions=Self>;
    unsafe fn uninitialized(length: usize) -> Self;
    unsafe fn set_unchecked(&mut self, i: usize, tuple: Self::Item);
}

A tuple of distributions: (Distribution<A>, Distribution<B>, ..)

Associated Types

type Item: Tuple<Distributions=Self>

A tuple that can be pushed/inserted into the tupled distributions

Required Methods

unsafe fn uninitialized(length: usize) -> Self

Creates a tuple of unitialized distributions, where each distribution has the same length

unsafe fn set_unchecked(&mut self, i: usize, tuple: Self::Item)

Writes each element of tuple into its corresponding distribution