Trait criterion_stats::tuple::TupledDistributionsBuilder

pub trait TupledDistributionsBuilder: Sized {
    type Item: Tuple<Builder = Self>;
    fn new(size: usize) -> Self;

    fn push(&mut self, tuple: Self::Item);

    fn extend(&mut self, other: &mut Self);

    fn complete(self) -> <Self::Item as Tuple>::Distributions;
}

A tuple of vecs used to build distributions.

Associated Types

type Item: Tuple<Builder = Self>

A tuple that can be pushed/inserted into the tupled distributions

Required Methods

fn new(size: usize) -> Self

Creates a new tuple of vecs

fn push(&mut self, tuple: Self::Item)

Push one element into each of the vecs

fn extend(&mut self, other: &mut Self)

Append one tuple of vecs to this one, leaving the vecs in the other tuple empty

fn complete(self) -> <Self::Item as Tuple>::Distributions

Convert the tuple of vectors into a tuple of distributions