Struct criterion_stats::univariate::kde::Kde

pub struct Kde<'a, A, K> where A: 'a + Floaty, K: Kernel<A> {
    // some fields omitted
}

Univariate kernel density estimator

Methods

impl<'a, A, K> Kde<'a, A, K> where A: 'a + Floaty, K: Kernel<A>

fn new(sample: &'a Sample<A>, k: K, bw: Bandwidth<A>) -> Kde<'a, A, K>

Creates a new kernel density estimator from the sample, using a kernel k and estimating the bandwidth using the method bw

fn bandwidth(&self) -> A

Returns the bandwidth used by the estimator

fn map(&self, xs: &[A]) -> Box<[A]>

Maps the KDE over xs

Multihreaded