Struct criterion_stats::univariate::outliers::tukey::LabeledSample

pub struct LabeledSample<'a, A> where A: 'a + Floaty {
    // some fields omitted
}

A classified/labeled sample.

The labeled data can be accessed using the indexing operator. The order of the data points is retained.

NOTE: Due to limitations in the indexing traits, only the label is returned. Once the IndexGet trait lands in stdlib, the indexing operation will return a (data_point, label) pair.

Methods

impl<'a, A> LabeledSample<'a, A> where A: Floaty

fn count(&self) -> (usize, usize, usize, usize, usize)

Returns the number of data points per label

Time: O(length)

fn fences(&self) -> (A, A, A, A)

Returns the fences used to classify the outliers

fn iter(&self) -> Iter<'a, A>

Returns an iterator over the labeled data