Struct criterion_stats::bivariate::Data

pub struct Data<'a, X, Y>(_, _)
 where
    X: 'a,
    Y: 'a;

Bivariate (X, Y) data

Invariants:

No NaNs in the data

At least two data points in the set

Methods

impl<'a, X, Y> Data<'a, X, Y>

pub fn len(&self) -> usize

Returns the length of the data set

pub fn is_empty(&self) -> bool

Checks whether the data set is empty

Important traits for Pairs<'a, X, Y>
impl<'a, X, Y> Iterator for Pairs<'a, X, Y> type Item = (&'a X, &'a Y);

pub fn iter(&self) -> Pairs<'a, X, Y>

Iterate over the data set

impl<'a, X, Y> Data<'a, X, Y> where
X: Float,
Y: Float,

pub fn new(xs: &'a [X], ys: &'a [Y]) -> Data<'a, X, Y>

Creates a new data set from two existing slices

pub fn bootstrap<T, S>(
&self,
nresamples: usize,
statistic: S
) -> T::Distributions where
S: Fn(Data<X, Y>) -> T,
S: Sync,
T: Tuple,
T::Distributions: Send,
T::Builder: Send,

Returns the bootstrap distributions of the parameters estimated by the statistic

Multi-threaded

Time: O(nresamples)

Memory: O(nresamples)

pub fn x(&self) -> &'a Sample<X>

Returns a view into the X data

pub fn y(&self) -> &'a Sample<Y>

Returns a view into the Y data