Struct criterion_stats::univariate::kde::Kde

pub struct Kde<'a, A, K> where
    A: 'a + Float,
    K: Kernel<A>,  { /* fields omitted */ }

Univariate kernel density estimator

Methods

impl<'a, A, K> Kde<'a, A, K> where
A: 'a + Float,
K: Kernel<A>,

pub fn new(sample: &'a Sample<A>, kernel: K, bw: Bandwidth<A>) -> Kde<'a, A, K>

Creates a new kernel density estimator from the sample, using a kernel and estimating the bandwidth using the method bw

pub fn bandwidth(&self) -> A

Returns the bandwidth used by the estimator

pub fn map(&self, xs: &[A]) -> Box<[A]>

Maps the KDE over xs

Multihreaded

pub fn estimate(&self, x: A) -> A

Estimates the probability density of x