Enum criterion_stats::univariate::outliers::tukey::Label

pub enum Label {
    HighMild,
    HighSevere,
    LowMild,
    LowSevere,
    NotAnOutlier,
}

Labels used to classify outliers

Variants
HighMild

A "mild" outlier in the "high" spectrum

HighSevere

A "severe" outlier in the "high" spectrum

LowMild

A "mild" outlier in the "low" spectrum

LowSevere

A "severe" outlier in the "low" spectrum

NotAnOutlier

A normal data point

Methods

impl Label

pub fn is_high(&self) -> bool

Checks if the data point has an "unusually" high value

pub fn is_mild(&self) -> bool

Checks if the data point is labeled as a "mild" outlier

pub fn is_low(&self) -> bool

Checks if the data point has an "unusually" low value

pub fn is_outlier(&self) -> bool

Checks if the data point is labeled as an outlier

pub fn is_severe(&self) -> bool

Checks if the data point is labeled as a "severe" outlier